+++ app/models/LicenseData.scala
@@ -2,6 +2,9 @@ package models
 
 import api.Permission
 
+import play.api.libs.json._
+
+
 /**
  * case class to handle specific license information. Currently attached to individual Datasets and Files.  
  */
@@ -41,5 +44,67 @@ case class LicenseData (
     def isRightsOwner(aName: String) = {
         m_rightsHolder == aName
     }
-}
 
+    /**
+     * Utility to return a url even if empty, but enough other attributes available to determine it
+     * this is repurposed from:
+     * function updateData(id, imageBase, sourceObject, authorName)  
+     * in updateLicenseInfo.js line:88
+     */
+    def urlViaAttributes() : String = {
+       if (m_licenseUrl != "") return m_licenseUrl
+       var licenseUrl = m_licenseUrl;
+       if (m_licenseType == "license2") {
+         //No checkboxes selected
+         if (!m_ccAllowCommercial && !m_ccAllowDerivative && !m_ccRequireShareAlike) {
+            licenseUrl = "http://creativecommons.org/licenses/by-nc-nd/3.0/";
+         }
+         //Only commercial selected
+         else if (m_ccAllowCommercial && !m_ccAllowDerivative && !m_ccRequireShareAlike) {
+            licenseUrl = "http://creativecommons.org/licenses/by-nd/3.0/";
+         }
+         //Only remixing selected
+         else if (!m_ccAllowCommercial && m_ccAllowDerivative && !m_ccRequireShareAlike) {
+            licenseUrl = "http://creativecommons.org/licenses/by-nc/3.0/";
+         }
+         //Remixing and Sharealike selected
+         else if (!m_ccAllowCommercial && m_ccAllowDerivative && m_ccRequireShareAlike) {
+            licenseUrl = "http://creativecommons.org/licenses/by-nc-sa/3.0/";
+         }
+         //All checkboxes selected
+         else if (m_ccAllowCommercial && m_ccAllowDerivative && m_ccRequireShareAlike) {
+            licenseUrl = "http://creativecommons.org/licenses/by-sa/3.0/";
+         }
+         //Commercial and Remixing selected
+         else if (m_ccAllowCommercial && m_ccAllowDerivative && !m_ccRequireShareAlike) {
+            licenseUrl = "http://creativecommons.org/licenses/by/3.0/";
+         }
+         //else { rightsHolder = 'Creative Commons';
+         //   licenseText = 'Specific level info'; }
+      }
+      else if (m_licenseType == "license3") {
+         licenseUrl = "http://creativecommons.org/publicdomain/zero/1.0/";
+      }
+      else {
+         licenseUrl = "https://dbpedia.org/page/All_rights_reserved";
+      }
+      return licenseUrl
+   }
+        
+   /**
+    * Utility function, similar to a json Write, to return string version in json-ld  format
+    * Should also return key
+    */
+    def to_jsonld () : JsValue = {
+       val licURI = this.urlViaAttributes()  //URI = URL except in one case:
+       val licURL = if (licURI != "https://dbpedia.org/page/All_rights_reserved") licURI
+                    else ""
+       val licLD = JsObject(Seq(
+                "@id" -> JsString(licURI),
+                "URL" -> JsString(licURL),
+                "@type" -> JsString("license"),
+                "Text" -> JsString(m_licenseText)  //added this DataType
+                ))
+       return licLD
+    }
+}